+++ src/tko/cmds/cmd_down.py
@@ -35,19 +35,19 @@ class CmdLineDown:
 
 
 class CmdDown:
+    test_case_filename = "cases.tio"
     def __init__(self, rep: Repository, task_key: str, settings: Settings):
         self.rep = rep
         self.task_key = task_key
         self.settings = settings
         self.game: Game | None = None
-        self.language: str | None = None
-        self.destiny_folder: str = ""
-        self.readme_path: str = ""
-        self.cache_url: str = ""
-        self.mapi_file: str = ""
-
+        self.destiny_folder = self.rep.get_task_folder_for_label(self.task_key)
+        self.readme_path =  os.path.join(self.destiny_folder, "Readme.md")
+        self.language: str = ""
+        self.actions = DownActions()
+        
     def set_fnprint(self, fnprint: Callable[[str], None]):
-        DownProblem.fnprint = fnprint
+        self.actions.fnprint = fnprint
         return self
 
     def set_game(self, game: Game):
@@ -64,28 +64,29 @@ class CmdDown:
 
     def download_readme(self, readme_remote_url: RemoteUrl) -> bool:
         try:
-            DownProblem.down_readme(self.readme_path, readme_remote_url)
+            self.actions.down_readme(self.readme_path, readme_remote_url)
             return True
         except urllib.error.HTTPError:
-            DownProblem.fnprint("falha: Arquivo de descrição não encontrado")
+            self.actions.send_to_print("falha: Arquivo de descrição não encontrado")
             return False
         except urllib.error.URLError:
-            DownProblem.fnprint("falha: Não consegui baixar a atividade, verifique sua internet")
+            self.actions.send_to_print("falha: Não consegui baixar a atividade, verifique sua internet")
             return False
 
-    def download_mapi(self) -> bool:
+    def download_mapi(self, cache_url: str) -> tuple[bool, str]:
         try:
-            mapi_url = self.cache_url + "mapi.json"
-            urllib.request.urlretrieve(mapi_url, self.mapi_file)
-            return True
+            mapi_url = cache_url + "mapi.json"
+            mapi_path = os.path.join(self.destiny_folder, "mapi.json")
+            urllib.request.urlretrieve(mapi_url, mapi_path)
+            return True, mapi_path
         except urllib.error.HTTPError:
-            return False
+            return False, ""
         except urllib.error.URLError:
-            DownProblem.fnprint("falha: não consegui baixar a atividade, verifique sua internet")
-            return False
+            self.actions.send_to_print("falha: não consegui baixar a atividade, verifique sua internet")
+            return False, ""
 
     def build_cases_from_readme(self, destiny_folder: str):
-        cases_tio_target = os.path.join(destiny_folder, "cases.tio")
+        cases_tio_target = os.path.join(destiny_folder, CmdDown.test_case_filename)
         param = Param.Manip()
         cb = CmdBuild(cases_tio_target, [self.readme_path], param)
         cb.set_quiet(True)
@@ -113,12 +114,13 @@ class CmdDown:
                 os.makedirs(drafts_folder, exist_ok=True)
                 break
             count += 1
-        DownProblem.fnprint("")
-        DownProblem.fnprint(f"Criando nova pasta de rascunhos: {os.path.basename(drafts_folder)}")
-        DownProblem.fnprint(f"Você precisa manualmente copiar os novos rascunhos")
+        self.actions.send_to_print("")
+        self.actions.send_to_print(f"Criando nova pasta de rascunhos: {os.path.basename(drafts_folder)}")
+        self.actions.send_to_print(f"Se quiser utilizar, copie manualmente")
+        self.actions.send_to_print(f"os novos rascunhos para a pasta {lang}")
         return drafts_folder
 
-    def check_and_remove_draft(self, drafts_folder: str, lang: str) -> None:
+    def remove_draft_folder_if_duplicated(self, drafts_folder: str, lang: str) -> tuple[bool, str]:
         destiny_folder = os.path.dirname(drafts_folder)
         parts = drafts_folder.split(".")
         if len(parts) > 1:
@@ -131,48 +133,71 @@ class CmdDown:
                     last_path = ".".join(parts[:-1]) + f".{last}"
                 if os.path.exists(last_path):
                     if self.folder_equals(drafts_folder, last_path):
-                        lastname = os.path.basename(last_path)
-                        newname = os.path.basename(drafts_folder)
-                        DownProblem.fnprint("")
-                        DownProblem.fnprint(f"Pasta {newname} igual a {lastname}, removendo {newname}")
                         shutil.rmtree(drafts_folder)
+                        return True, last_path
             except ValueError:
                 pass
+        return False, ""
+
+
+    def download_readme_cases_and_drafts(self, mapi_path: str) -> None:
+        lang = self.check_and_select_language()
+
+        loaded_json: dict[str, Any] | None = None
+        if mapi_path != "" and os.path.isfile(mapi_path):
+            encoding = Decoder.get_encoding(mapi_path)
+            with open(mapi_path, encoding=encoding) as f:
+                loaded_json = json.load(f)
+            if loaded_json is not None:
+                self.actions.unpack_problem_files(loaded_json, self.destiny_folder)
+
+        test_case_path = os.path.join(self.destiny_folder, CmdDown.test_case_filename)
+        if not os.path.isfile(test_case_path):
+            self.build_cases_from_readme(self.destiny_folder)
+        with open(test_case_path, "r", encoding="utf-8") as f:
+            content = f.read()
+        if content.strip() == "":
+            os.remove(test_case_path)
+        self.actions.cached = True
+        drafts_folder = self.find_folder_for_drafts(self.destiny_folder, lang)
+        if not self.actions.unpack_json_drafts(loaded_json, drafts_folder, lang):
+            self.actions.create_default_draft(drafts_folder, lang)
+        
+        removed, last_path = self.remove_draft_folder_if_duplicated(drafts_folder, lang)
+        self.actions.cached = False
+        if not removed:
+            for msg in self.actions.cache_msgs:
+                self.actions.send_to_print(msg)
+            self.actions.cache_msgs.clear()
+        else:
+            self.actions.send_to_print(f"Último rascunho em {os.path.basename(last_path)}")
+        self.actions.send_to_print("")
+        self.actions.send_to_print("Atividade baixada com sucesso")
 
     def download_from_url(self, task_source: str) -> bool:
         readme_remote_url = RemoteUrl(task_source)
         remote_url = readme_remote_url.get_raw_url()
-        self.cache_url = os.path.dirname(remote_url) + "/.cache/"
-        self.destiny_folder = self.rep.get_task_folder_for_label(self.task_key)
-        self.readme_path =  os.path.join(self.destiny_folder, "Readme.md")
-        self.mapi_file = os.path.join(self.destiny_folder, "mapi.json")
-        # backup_folder = DownProblem.backup_and_create_problem_folder(self.destiny_folder)
         os.makedirs(self.destiny_folder, exist_ok=True)
         if not self.download_readme(readme_remote_url):
             self.remove_empty_destiny_folder()
             return False
-        
-        lang = self.check_and_select_language()
-        drafts_done: bool = False
-        if self.download_mapi():
-            encoding = Decoder.get_encoding(self.mapi_file)
-            with open(self.mapi_file, encoding=encoding) as f:
-                loaded_json = json.load(f)
-            os.remove(self.mapi_file)
-            DownProblem.unpack_problem_files(loaded_json, self.destiny_folder)
-            drafts_folder = self.find_folder_for_drafts(self.destiny_folder, lang)
-            if DownProblem.unpack_json_drafts(loaded_json, drafts_folder, lang):
-                drafts_done = True
-        else:
-            drafts_folder = self.find_folder_for_drafts(self.destiny_folder, lang)
-            self.build_cases_from_readme(self.destiny_folder)
-
-        if not drafts_done:
-            DownProblem.create_default_draft(drafts_folder, lang)
-        self.check_and_remove_draft(drafts_folder, lang)
-        DownProblem.fnprint("")
-        DownProblem.fnprint("Atividade baixada com sucesso")
+        cache_url = os.path.dirname(remote_url) + "/.cache/"
+        _, mapi_path = self.download_mapi(cache_url)
+        self.download_readme_cases_and_drafts(mapi_path)
+        if os.path.isfile(mapi_path):
+            os.remove(mapi_path)
         return True
+
+    def download_from_external_file(self, task_source: str) -> bool:
+        if not os.path.exists(task_source):
+            self.actions.send_to_print(f"Arquivo fonte não encontrado para carregar a atividade {self.task_key}")
+            return False
+        os.makedirs(self.destiny_folder, exist_ok=True)
+        self.actions.compare_and_save(Decoder.load(task_source), self.readme_path)
+        mapi_file = os.path.join(os.path.dirname(task_source), ".cache", "mapi.json")
+        self.download_readme_cases_and_drafts(mapi_file)
+        return True
+
 
     @staticmethod
     def folder_equals(folder1: str, folder2: str) -> bool:
@@ -192,33 +217,6 @@ class CmdDown:
         return True
 
 
-    def download_from_external_file(self, task_source: str) -> bool:
-        if not os.path.exists(task_source):
-            DownProblem.fnprint("Arquivo fonte não encontrado")
-        self.cache_url = os.path.join(os.path.dirname(task_source), ".cache")
-        self.destiny_folder = self.rep.get_task_folder_for_label(self.task_key)
-        self.readme_path =  os.path.join(self.destiny_folder, "Readme.md")
-        self.mapi_file = os.path.join(self.cache_url, "mapi.json")
-        backup_folder = DownProblem.backup_and_create_problem_folder(self.destiny_folder)
-        # Decoder.save(self.readme_path, Decoder.load(task_source))
-        DownProblem.compare_and_save(Decoder.load(task_source), self.readme_path)
-
-        lang = self.check_and_select_language()
-        
-        if os.path.isfile(self.mapi_file):
-            encoding = Decoder.get_encoding(self.mapi_file)
-            with open(self.mapi_file, encoding=encoding) as f:
-                loaded_json = json.load(f)
-            DownProblem.unpack_problem_files(loaded_json, self.destiny_folder)
-            if not DownProblem.unpack_json_drafts(loaded_json, self.destiny_folder, lang):
-                if not backup_folder:
-                    DownProblem.create_default_draft(self.destiny_folder, lang)
-        else:
-            self.build_cases_from_readme(self.destiny_folder)
-            DownProblem.create_default_draft(self.destiny_folder, lang)
-        DownProblem.fnprint("")
-        DownProblem.fnprint("Atividade carregada com sucesso")
-        return True
 
     def execute(self) -> bool:
         task = self.rep.game.get_task(self.task_key)
@@ -226,13 +224,13 @@ class CmdDown:
             return self.download_from_url(task.link)
         if task.link_type == Task.Types.IMPORT_FILE:
             return self.download_from_external_file(task.link)
-        DownProblem.fnprint("falha: link para atividade não possui link para download")
+        self.actions.send_to_print("falha: link para atividade não possui link para download")
         return False
 
     def check_and_select_language(self) -> str:
         language_def = self.rep.get_lang()
 
-        if self.language is None:
+        if self.language == "":
             if language_def != "":
                 self.language = language_def
             else:
@@ -240,42 +238,51 @@ class CmdDown:
                 self.language = input()
         return self.language
 
-class DownProblem:
-    fnprint: Callable[[str], None] = print
+class DownActions:
 
-    @staticmethod
-    def __create_file(content: str, path: str, label: str="", parts: int = 2):
+    def __init__(self):
+        self.fnprint: Callable[[str], None] = print
+        self.cache_msgs: list[str] = []
+        self.cached: bool = False
+
+    def send_to_print(self, text: str):
+        if self.cached:
+            self.cache_msgs.append(text)
+        else:
+            self.fnprint(text)
+
+    def __create_file(self, content: str, path: str, label: str="", parts: int = 2):
         Decoder.save(path, content)
-        DownProblem.fnprint(DownProblem.folder_and_file(path, parts) + " " + label)
+        self.send_to_print(DownActions.folder_and_file(path, parts) + " " + label)
 
-    @staticmethod
-    def unpack_json_drafts(loaded: dict[str, Any], destiny: str, lang: str) -> bool:
+    def unpack_json_drafts(self, loaded: dict[str, Any] | None, destiny: str, lang: str) -> bool:
+        if loaded is None:
+            return False
         found = False
         if "draft" in loaded:
             if lang in loaded["draft"]:
                 for file in loaded["draft"][lang]:
                     path = os.path.join(destiny, file["name"])
-                    DownProblem.__create_file(file["contents"], path, "(Rascunho)", 3)
+                    self.__create_file(file["contents"], path, "(Rascunho)", 3)
                     found = True
 
         return found
 
-    @staticmethod
-    def unpack_problem_files(loaded: dict[str, Any], destiny: str):
+    def unpack_problem_files(self, loaded: dict[str, Any], destiny: str):
         if "upload" in loaded:
             for file in loaded["upload"]:
                 name = file["name"]
                 if name == "vpl_evaluate.cases":
-                    name = "cases.tio"
+                    name = CmdDown.test_case_filename
                 path = os.path.join(destiny, name)
-                DownProblem.compare_and_save(file["contents"], path)
+                self.compare_and_save(file["contents"], path)
         if "keep" in loaded:
             for file in loaded["keep"]:
                 name = file["name"]
                 if name == "vpl_evaluate.cases":
-                    name = "cases.tio"
+                    name = CmdDown.test_case_filename
                 path = os.path.join(destiny, name)
-                DownProblem.compare_and_save(file["contents"], path)
+                self.compare_and_save(file["contents"], path)
 
     @staticmethod
     def folder_and_file(path: str, parts: int = 2) -> str:
@@ -284,52 +291,25 @@ class DownProblem:
         pieces = pieces[-parts:]  # Get the last 'parts' elements
         return os.path.join(*pieces)
 
-    @staticmethod
-    def  compare_and_save(content: str, path: str):
+    def compare_and_save(self, content: str, path: str):
         if not os.path.exists(path):
             Decoder.save(path, content)
-            DownProblem.fnprint(DownProblem.folder_and_file(path) + " (Novo)")
+            self.send_to_print(self.folder_and_file(path) + " (Novo)")
         else:
             path_content = Decoder.load(path)
             if path_content != content:
-                DownProblem.fnprint(DownProblem.folder_and_file(path) + " (Atualizado)")
+                self.send_to_print(self.folder_and_file(path) + " (Atualizado)")
                 Decoder.save(path, content)
             else:
-                DownProblem.fnprint(DownProblem.folder_and_file(path) + " (Inalterado)")
+                self.send_to_print(self.folder_and_file(path) + " (Inalterado)")
 
-    @staticmethod
-    def down_readme(readme_path: str,  remote_url: RemoteUrl):
+    def down_readme(self, readme_path: str,  remote_url: RemoteUrl):
         temp_file = tempfile.mktemp()
         remote_url.download_absolute_to(temp_file)
         content = Decoder.load(temp_file)
-        DownProblem.compare_and_save(content, readme_path)
-    
-    @staticmethod
-    def backup_and_create_problem_folder(destiny: str) -> str:
-        """ If destiny exists, rename it to destiny+1, destiny+2, etc.
-        If backup was created, return the new name.
-        If not, return empty string.
-        """
-        # new_destiny = ""
-        # if os.path.exists(destiny): # move folder to 
-        #     count = 1
-        #     while True:
-        #         new_destiny = "{}+{}".format(destiny, count)
-        #         if not os.path.exists(new_destiny):
-        #             shutil.move(destiny, new_destiny)
-        #             DownProblem.fnprint("Pasta {} já existe".format(os.path.basename(destiny)))
-        #             DownProblem.fnprint("Renomeando para {}".format(os.path.basename(new_destiny)))
-        #             break
-        #         else:
-        #             count += 1
-        # nova = "" if new_destiny == "" else " nova"
-        # DownProblem.fnprint("Criando{} pasta {}".format(nova, os.path.basename(destiny)))
-        os.makedirs(destiny, exist_ok=True)
-        # DownProblem.fnprint("")
-        return ""
+        self.compare_and_save(content, readme_path)
 
-    @staticmethod
-    def create_default_draft(destiny: str, language: str):
+    def create_default_draft(self, destiny: str, language: str):
         filename = "draft."
         draft_path = os.path.join(destiny, filename + language)
         os.makedirs(os.path.dirname(draft_path), exist_ok=True)
@@ -339,6 +319,6 @@ class DownProblem:
                     f.write(Drafts.drafts[language])
                 else:
                     f.write("")
-            DownProblem.fnprint(DownProblem.folder_and_file(draft_path, 3) + " (Vazio)")
+            self.send_to_print(self.folder_and_file(draft_path, 3) + " (Vazio)")
         else:
-            DownProblem.fnprint(DownProblem.folder_and_file(draft_path, 3) + " (Não sobrescrito)")
+            self.send_to_print(self.folder_and_file(draft_path, 3) + " (Não sobrescrito)")